@@ -1,11 +1,11 @@
 apply plugin: 'com.android.library'
+
 android {
-    compileSdkVersion 24
-    buildToolsVersion "26.0.2"
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 9
-        targetSdkVersion 24
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -15,16 +15,16 @@ android {
         }
     }
 }
-def name = 'GenericAdapter_112.jar'
+def name = 'GenericAdapter_200.aar'
 def toPath = '../librarys/'
 task clearAar(type: Delete) {
     delete toPath+'/'+name
 }
 task makeAar(type: Copy) {
-    from('build/intermediates/bundles/release/')
+    from('build/outputs/aar/')
     into(toPath)
-    include('classes.jar')
-    rename('classes.jar', name)
+    include('genericadapter-release.aar')
+    rename('genericadapter-release.aar', name)
 }
 makeAar.dependsOn(clearAar, build)
 
