@@ -1,7 +1,8 @@
 package app.neetoffice.com.genericadaptersample;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by Deo on 2016/3/16.
