@@ -3,13 +3,8 @@ package app.neetoffice.com.genericadaptersample;
 import android.content.Context;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.util.Log;
 import android.view.MenuItem;
 import android.view.ViewGroup;
 import android.widget.EditText;
@@ -20,6 +15,10 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import jp.wasabeef.recyclerview.animators.FadeInUpAnimator;
 import library.neetoffice.com.genericadapter.CellView;
 import library.neetoffice.com.genericadapter.GenericStickyHeaderRecyclerAdapter;
