@@ -1,10 +1,7 @@
 package library.neetoffice.com.genericadapter;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 
+import androidx.recyclerview.widget.RecyclerView;
 import library.neetoffice.com.genericadapter.base.OnScrollCallBack;
 
 /**
